+++ crates/miden-tx/tests/integration/scripts/swap.rs
@@ -85,7 +85,7 @@ fn prove_consume_swap_note() {
 
     let target_account = mock_chain.add_existing_wallet(Auth::BasicAuth, vec![requested_asset]);
     mock_chain.add_pending_note(note.clone());
-    mock_chain.seal_block(None);
+    mock_chain.seal_next_block();
 
     let consume_swap_note_tx = mock_chain
         .build_tx_context(target_account.id(), &[note.id()], &[])